+++ x/auth/tx/generator.go
@@ -4,14 +4,13 @@ import (
 	"fmt"
 
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/codec"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
 
 type generator struct {
-	marshaler   codec.Marshaler
 	pubkeyCodec types.PublicKeyCodec
 	handler     signing.SignModeHandler
 	decoder     sdk.TxDecoder
@@ -21,20 +20,19 @@ type generator struct {
 }
 
 // NewTxConfig returns a new protobuf TxConfig using the provided Marshaler, PublicKeyCodec and SignModeHandler.
-func NewTxConfig(marshaler codec.Marshaler, pubkeyCodec types.PublicKeyCodec, signModeHandler signing.SignModeHandler) client.TxConfig {
+func NewTxConfig(anyUnpacker codectypes.AnyUnpacker, pubkeyCodec types.PublicKeyCodec, signModeHandler signing.SignModeHandler) client.TxConfig {
 	return &generator{
-		marshaler:   marshaler,
 		pubkeyCodec: pubkeyCodec,
 		handler:     signModeHandler,
-		decoder:     DefaultTxDecoder(marshaler, pubkeyCodec),
-		encoder:     DefaultTxEncoder(marshaler),
-		jsonDecoder: DefaultJSONTxDecoder(marshaler, pubkeyCodec),
-		jsonEncoder: DefaultJSONTxEncoder(marshaler),
+		decoder:     DefaultTxDecoder(anyUnpacker, pubkeyCodec),
+		encoder:     DefaultTxEncoder(),
+		jsonDecoder: DefaultJSONTxDecoder(anyUnpacker, pubkeyCodec),
+		jsonEncoder: DefaultJSONTxEncoder(),
 	}
 }
 
 func (g generator) NewTxBuilder() client.TxBuilder {
-	return newBuilder(g.marshaler, g.pubkeyCodec)
+	return newBuilder(g.pubkeyCodec)
 }
 
 // WrapTxBuilder returns a builder from provided transaction